backend/django/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-rest-auth==2.1.3
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-redis==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
drf-spectacular==0.14.0
drf-writable-nested==0.6.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
lxml==4.6.3
Markdown==3.3.4
mccabe==0.6.1
mysql-connector-python==8.0.24
mysqlclient==2.0.3
oauthlib==3.1.0
phonenumbers==8.12.21
Pillow==8.1.2
protobuf==3.15.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.1
PyJWT==2.0.1
pyrsistent==0.17.3
python-docx==0.8.11
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
